changed uart send

+++ IoTSense/Inc/si1145.h
@@ -11,9 +11,9 @@
 #include "stm32f0xx_hal.h"
 
 void si1145_init();
-uint16_t si1145_getVisible();
-uint16_t si1145_getIR();
-uint16_t si1145_getUVIndex();
+uint32_t si1145_getVisible();
+uint32_t si1145_getIR();
+uint32_t si1145_getUVIndex();
 uint8_t si1145_getResponse();
 void si1145_sendNOP();
 char* si1145_getStatus();
